+++ src/main/kotlin/one/wabbit/typeclass/plugin/TypeclassCompilerPluginRegistrar.kt
@@ -12,7 +12,10 @@ internal class TypeclassCompilerPluginRegistrar : CompilerPluginRegistrar() {
     override val supportsK2: Boolean = true
 
     override fun ExtensionStorage.registerExtensions(configuration: CompilerConfiguration) {
-        val sharedState = TypeclassPluginSharedState(configuration.toTypeclassConfiguration())
+        val sharedState =
+            TypeclassPluginSharedState(
+                configuration = configuration.toTypeclassConfiguration(),
+            )
         FirExtensionRegistrarAdapter.registerExtension(TypeclassFirExtensionRegistrar(sharedState))
         IrGenerationExtension.Companion.registerExtension(TypeclassIrGenerationExtension(sharedState))
     }